some update enhancements

+++ electron/Updates.js
@@ -3,35 +3,44 @@ const path = require("path");
 const fs = require("fs/promises");
 const fsSync = require("fs");
 const { defaultUpdatePath } = require("./constants");
+const { platform } = require("os");
+
+const supportedUpdatePlatforms = ["win32"];
 
 async function setupUpdates(updateData) {
-  let updateTo;
+  if (platform === "win32") {
+    let updateTo;
 
-  try {
-    updateTo = fs.readFile(path.join(app.getPath("userData"), "updatedir.txt"));
-  } catch {}
+    try {
+      updateTo = fs.readFile(
+        path.join(app.getPath("userData"), "updatedir.txt")
+      );
+    } catch {}
 
-  const updateExe = path.join(updateTo, "update/installer.exe");
+    const updateExe = path.join(updateTo, "update/installer.exe");
 
-  if (fsSync.existsSync(updateExe)) fs.unlink(updateExe);
+    if (fsSync.existsSync(updateExe)) fs.unlink(updateExe);
 
-  fs.copyFile(path.join(__dirname, "../update/installer.exe"), updateExe);
+    fs.copyFile(path.join(__dirname, "../update/installer.exe"), updateExe);
 
-  spawnSync(
-    updateExe,
-    [
-      updateData.updateUrl,
-      updateTo || resolveToAbsolutePath(defaultUpdatePath),
-      updateData.name,
-      path.join(__dirname, "../../.."),
-    ],
-    {
-      detached: true,
-    }
-  );
+    spawnSync(
+      updateExe,
+      [
+        updateData.updateUrl,
+        updateTo || resolveToAbsolutePath(defaultUpdatePath),
+        updateData.name,
+        path.join(__dirname, "../../.."),
+      ],
+      {
+        detached: true,
+      }
+    );
+  } else return false;
 }
 
 function checkUpdates(elFeed) {
+  if (!supportedUpdatePlatforms.includes(platform)) return undefined;
+
   return fetch(elFeed)
     .then((res) => {
       if (res.status === 204) {
